+++ USTrails/USTrails.API/Models/DTO/StateDto.cs
@@ -2,7 +2,7 @@
 {
     public class StateDto
     {
-        public byte Id { get; set; }
+        public short Id { get; set; }
         public string Code { get; set; }
         public string Name { get; set; }
         public string? StateImageUrl { get; set; }
